+++ R/expr__expr.R
@@ -1289,18 +1289,19 @@ Expr_rechunk = "use_extendr_wrapper"
 #' @param reverse bool, default FALSE, if true roll over vector from back to forth
 #' @return Expr
 #' @aliases Expr_cumsum
-#' @name Expr_cumsum
+#' @name Expr_cum_sum
 #' @details
 #' The Dtypes Int8, UInt8, Int16 and UInt16 are cast to
 #' Int64 before summing to prevent overflow issues.
 #' @format NULL
 #' @examples
 #' pl$DataFrame(list(a = 1:4))$select(
-#'   pl$col("a")$cumsum()$alias("cumsum"),
-#'   pl$col("a")$cumsum(reverse = TRUE)$alias("cumsum_reversed")
+#'   pl$col("a")$cum_sum()$alias("cum_sum"),
+#'   pl$col("a")$cum_sum(reverse = TRUE)$alias("cum_sum_reversed")
 #' )
-Expr_cumsum = function(reverse = FALSE) {
-  .pr$Expr$cumsum(self, reverse)
+Expr_cum_sum = function(reverse = FALSE) {
+  .pr$Expr$cum_sum(self, reverse) |>
+    unwrap("in cum_sum():")
 }
 
 
@@ -1309,8 +1310,8 @@ Expr_cumsum = function(reverse = FALSE) {
 #' @keywords Expr
 #' @param reverse bool, default FALSE, if true roll over vector from back to forth
 #' @return Expr
-#' @aliases cumprod
-#' @name Expr_cumprod
+#' @aliases cum_prod
+#' @name Expr_cum_prod
 #' @details
 #' The Dtypes Int8, UInt8, Int16 and UInt16 are cast to
 #' Int64 before summing to prevent overflow issues.
@@ -1318,11 +1319,12 @@ Expr_cumsum = function(reverse = FALSE) {
 #' @format NULL
 #' @examples
 #' pl$DataFrame(list(a = 1:4))$select(
-#'   pl$col("a")$cumprod()$alias("cumprod"),
-#'   pl$col("a")$cumprod(reverse = TRUE)$alias("cumprod_reversed")
+#'   pl$col("a")$cum_prod()$alias("cum_prod"),
+#'   pl$col("a")$cum_prod(reverse = TRUE)$alias("cum_prod_reversed")
 #' )
-Expr_cumprod = function(reverse = FALSE) {
-  .pr$Expr$cumprod(self, reverse)
+Expr_cum_prod = function(reverse = FALSE) {
+  .pr$Expr$cum_prod(self, reverse) |>
+    unwrap("in cum_prod():")
 }
 
 #' Cumulative minimum
@@ -1330,8 +1332,8 @@ Expr_cumprod = function(reverse = FALSE) {
 #' @keywords Expr
 #' @param reverse bool, default FALSE, if true roll over vector from back to forth
 #' @return Expr
-#' @aliases cummin
-#' @name Expr_cummin
+#' @aliases cum_min
+#' @name Expr_cum_min
 #' @details
 #' The Dtypes Int8, UInt8, Int16 and UInt16 are cast to
 #' Int64 before summing to prevent overflow issues.
@@ -1340,11 +1342,12 @@ Expr_cumprod = function(reverse = FALSE) {
 #' @format NULL
 #' @examples
 #' pl$DataFrame(list(a = 1:4))$select(
-#'   pl$col("a")$cummin()$alias("cummin"),
-#'   pl$col("a")$cummin(reverse = TRUE)$alias("cummin_reversed")
+#'   pl$col("a")$cum_min()$alias("cum_min"),
+#'   pl$col("a")$cum_min(reverse = TRUE)$alias("cum_min_reversed")
 #' )
-Expr_cummin = function(reverse = FALSE) {
-  .pr$Expr$cummin(self, reverse)
+Expr_cum_min = function(reverse = FALSE) {
+  .pr$Expr$cum_min(self, reverse) |>
+    unwrap("in cum_min():")
 }
 
 #' Cumulative maximum
@@ -1353,7 +1356,7 @@ Expr_cummin = function(reverse = FALSE) {
 #' @param reverse bool, default FALSE, if true roll over vector from back to forth
 #' @return Expr
 #' @aliases cummin
-#' @name Expr_cummin
+#' @name Expr_cum_max
 #' @details
 #' The Dtypes Int8, UInt8, Int16 and UInt16 are cast to
 #' Int64 before summing to prevent overflow issues.
@@ -1362,11 +1365,12 @@ Expr_cummin = function(reverse = FALSE) {
 #' @format NULL
 #' @examples
 #' pl$DataFrame(list(a = 1:4))$select(
-#'   pl$col("a")$cummax()$alias("cummux"),
-#'   pl$col("a")$cummax(reverse = TRUE)$alias("cummax_reversed")
+#'   pl$col("a")$cum_max()$alias("cummux"),
+#'   pl$col("a")$cum_max(reverse = TRUE)$alias("cum_max_reversed")
 #' )
-Expr_cummax = function(reverse = FALSE) {
-  .pr$Expr$cummax(self, reverse)
+Expr_cum_max = function(reverse = FALSE) {
+  .pr$Expr$cum_max(self, reverse) |>
+    unwrap("in cum_max():")
 }
 
 #' Cumulative count
@@ -1375,22 +1379,23 @@ Expr_cummax = function(reverse = FALSE) {
 #' @keywords Expr
 #' @param reverse bool, default FALSE, if true roll over vector from back to forth
 #' @return Expr
-#' @aliases cumcount
-#' @name Expr_cumcount
+#' @aliases cum_count
+#' @name Expr_cum_count
 #' @details
 #' The Dtypes Int8, UInt8, Int16 and UInt16 are cast to
 #' Int64 before summing to prevent overflow issues.
 #'
-#' cumcount does not seem to count within lists.
+#' cum_count does not seem to count within lists.
 #'
 #' @format NULL
 #' @examples
 #' pl$DataFrame(list(a = 1:4))$select(
-#'   pl$col("a")$cumcount()$alias("cumcount"),
-#'   pl$col("a")$cumcount(reverse = TRUE)$alias("cumcount_reversed")
+#'   pl$col("a")$cum_count()$alias("cum_count"),
+#'   pl$col("a")$cum_count(reverse = TRUE)$alias("cum_count_reversed")
 #' )
-Expr_cumcount = function(reverse = FALSE) {
-  .pr$Expr$cumcount(self, reverse)
+Expr_cum_count = function(reverse = FALSE) {
+  .pr$Expr$cum_count(self, reverse) |>
+    unwrap("in cum_count():")
 }
 
 
@@ -1721,16 +1726,17 @@ Expr_sort_by = function(by, descending = FALSE) {
 #' @return Expr
 #' @keywords Expr
 #' @aliases take
-#' @name Expr_take
+#' @name Expr_gather
 #' @details
 #' similar to R indexing syntax e.g. `letters[c(1,3,5)]`, however as an expression, not as eager computation
 #' exceeding
 #'
 #' @format NULL
 #' @examples
-#' pl$select(pl$lit(0:10)$take(c(1, 8, 0, 7)))
-Expr_take = function(indices) {
-  .pr$Expr$take(self, pl$lit(indices))
+#' pl$select(pl$lit(0:10)$gather(c(1, 8, 0, 7)))
+Expr_gather = function(indices) {
+  .pr$Expr$gather(self, pl$lit(indices)) |>
+    unwrap("in $gather():")
 }
 
 
@@ -1751,7 +1757,8 @@ Expr_take = function(indices) {
 #'   pl$lit(0:3)$shift(2)$alias("shift+2")
 #' )
 Expr_shift = function(periods = 1) {
-  .pr$Expr$shift(self, periods)
+  .pr$Expr$shift(self, periods) |>
+    unwrap("in $shift():")
 }
 
 #' Shift and fill values
@@ -1773,7 +1780,8 @@ Expr_shift = function(periods = 1) {
 #'   pl$lit(0:3)$shift_and_fill(2, fill_value = pl$lit(42) / 2)$alias("shift+2")
 #' )
 Expr_shift_and_fill = function(periods, fill_value) {
-  .pr$Expr$shift_and_fill(self, periods, pl$lit(fill_value))
+  .pr$Expr$shift_and_fill(self, periods, pl$lit(fill_value)) |>
+    unwrap("in $shift_and_fill():")
 }
 
 
@@ -1806,7 +1814,7 @@ Expr_fill_null = function(value = NULL, strategy = NULL, limit = NULL) {
     ),
 
     # the two use cases
-    !is.null(value), .pr$Expr$fill_null(self, pl$lit(value)),
+    !is.null(value), unwrap(.pr$Expr$fill_null(self, value)),
     is.null(value), unwrap(.pr$Expr$fill_null_with_strategy(self, strategy, limit)),
 
     # catch failed any match
@@ -2324,7 +2332,7 @@ Expr_where = Expr_filter
 #' explode/flatten does not support categorical
 #'
 #' @examples
-#' pl$DataFrame(list(a = letters))$select(pl$col("a")$explode()$take(0:5))
+#' pl$DataFrame(list(a = letters))$select(pl$col("a")$explode()$gather(0:5))
 #'
 #' listed_group_df = pl$DataFrame(iris[c(1:3, 51:53), ])$group_by("Species")$agg(pl$all())
 #' print(listed_group_df)
@@ -2357,9 +2365,9 @@ Expr_flatten = "use_extendr_wrapper"
 #' @format NULL
 #'
 #' @examples
-#' pl$DataFrame(list(a = 0:24))$select(pl$col("a")$take_every(6))
-Expr_take_every = function(n) {
-  unwrap(.pr$Expr$take_every(self, n))
+#' pl$DataFrame(list(a = 0:24))$select(pl$col("a")$gather_every(6))
+Expr_gather_every = function(n) {
+  unwrap(.pr$Expr$gather_every(self, n))
 }