+++ src/rust/src/rlib.rs
@@ -14,27 +14,47 @@ use std::result::Result;
 
 #[extendr]
 fn min_horizontal(dotdotdot: Robj) -> RResult<Expr> {
-    Ok(polars::lazy::dsl::min_horizontal(robj_to!(VecPLExprCol, dotdotdot)?).into())
+    Ok(
+        polars::lazy::dsl::min_horizontal(robj_to!(VecPLExprCol, dotdotdot)?)
+            .map_err(polars_to_rpolars_err)?
+            .into(),
+    )
 }
 
 #[extendr]
 fn max_horizontal(dotdotdot: Robj) -> RResult<Expr> {
-    Ok(polars::lazy::dsl::max_horizontal(robj_to!(VecPLExprCol, dotdotdot)?).into())
+    Ok(
+        polars::lazy::dsl::max_horizontal(robj_to!(VecPLExprCol, dotdotdot)?)
+            .map_err(polars_to_rpolars_err)?
+            .into(),
+    )
 }
 
 #[extendr]
 fn sum_horizontal(dotdotdot: Robj) -> RResult<Expr> {
-    Ok(polars::lazy::dsl::sum_horizontal(robj_to!(VecPLExprCol, dotdotdot)?).into())
+    Ok(
+        polars::lazy::dsl::sum_horizontal(robj_to!(VecPLExprCol, dotdotdot)?)
+            .map_err(polars_to_rpolars_err)?
+            .into(),
+    )
 }
 
 #[extendr]
 fn all_horizontal(dotdotdot: Robj) -> RResult<Expr> {
-    Ok(polars::lazy::dsl::all_horizontal(robj_to!(VecPLExprCol, dotdotdot)?).into())
+    Ok(
+        polars::lazy::dsl::all_horizontal(robj_to!(VecPLExprCol, dotdotdot)?)
+            .map_err(polars_to_rpolars_err)?
+            .into(),
+    )
 }
 
 #[extendr]
 fn any_horizontal(dotdotdot: Robj) -> RResult<Expr> {
-    Ok(polars::lazy::dsl::any_horizontal(robj_to!(VecPLExprCol, dotdotdot)?).into())
+    Ok(
+        polars::lazy::dsl::any_horizontal(robj_to!(VecPLExprCol, dotdotdot)?)
+            .map_err(polars_to_rpolars_err)?
+            .into(),
+    )
 }
 
 #[extendr]
@@ -284,7 +304,6 @@ extendr_module! {
     fn max_horizontal;
     fn sum_horizontal;
     fn coalesce_exprs;
-    fn sum_exprs;
 
     fn concat_list;
     fn concat_str;